refactor(app): getLabwareLocationCombos -> getLegacyLabwareLocationCombos
