+++ app/src/organisms/LegacyLabwarePositionCheck/utils/getProbeBasedLPCSteps.ts
@@ -1,7 +1,7 @@
 import { isEqual } from 'lodash'
 import { SECTIONS } from '../constants'
 import { getLabwareDefURI, getPipetteNameSpecs } from '@opentrons/shared-data'
-import { getLabwareLocationCombos } from '/app/organisms/LegacyApplyHistoricOffsets/hooks/getLabwareLocationCombos'
+import { getLegacyLabwareLocationCombos } from '/app/src/organisms/LegacyApplyHistoricOffsets/hooks/getLegacyLabwareLocationCombos'
 import { getLabwareDefinitionsFromCommands } from '@opentrons/components'
 
 import type {
@@ -9,7 +9,7 @@ import type {
   LoadedPipette,
 } from '@opentrons/shared-data'
 import type { LabwarePositionCheckStep, CheckPositionsStep } from '../types'
-import type { LabwareLocationCombo } from '/app/organisms/LegacyApplyHistoricOffsets/hooks/getLabwareLocationCombos'
+import type { LabwareLocationCombo } from '/app/src/organisms/LegacyApplyHistoricOffsets/hooks/getLegacyLabwareLocationCombos'
 
 function getPrimaryPipetteId(pipettes: LoadedPipette[]): string {
   if (pipettes.length < 1) {
@@ -47,7 +47,7 @@ function getAllCheckSectionSteps(
   protocolData: CompletedProtocolAnalysis
 ): CheckPositionsStep[] {
   const { pipettes, commands, labware, modules = [] } = protocolData
-  const labwareLocationCombos = getLabwareLocationCombos(
+  const labwareLocationCombos = getLegacyLabwareLocationCombos(
     commands,
     labware,
     modules
